+++ NEWS.md
@@ -1,3 +1,7 @@
 # mzrtools.dev 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added `mz_atomize` function to parse molecular formulas.
+* Added `mz_iso_list` to enumerate possible isotope combinations.
+* Added `mz_iso_probs` to calculate isotope probabilities.
+* Added `mz_calculate` to calculate molecular masses.